add detail/value_t.hpp
